[WEBKIT2] Fix for compilation warnings in WebContext.cpp https://bugs.webkit.org/show_bug.cgi?id=69080
Fix compilation warning in WebKit2 builds.

Patch by Goutham J <gouthamj@motorola.com> on 2011-12-01
Reviewed by Hajime Morita.

* UIProcess/WebContext.cpp:
(WebKit::WebContext::processDidFinishLaunching):
(WebKit::WebContext::startMemorySampler):

Canonical link: https://commits.webkit.org/90217@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@101746 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebKit2/UIProcess/WebContext.cpp
@@ -346,7 +346,7 @@ void WebContext::processDidFinishLaunching(WebProcessProxy* process)
     if (m_memorySamplerEnabled) {
         SandboxExtension::Handle sampleLogSandboxHandle;        
         double now = WTF::currentTime();
-        String sampleLogFilePath = String::format("WebProcess%llu", static_cast<uint64_t>(now));
+        String sampleLogFilePath = String::format("WebProcess%llu", static_cast<unsigned long long>(now));
         sampleLogFilePath = SandboxExtension::createHandleForTemporaryFile(sampleLogFilePath, SandboxExtension::WriteOnly, sampleLogSandboxHandle);
         
         m_process->send(Messages::WebProcess::StartMemorySampler(sampleLogSandboxHandle, sampleLogFilePath, m_memorySamplerInterval), 0);
@@ -787,7 +787,7 @@ void WebContext::startMemorySampler(const double interval)
     // For WebProcess
     SandboxExtension::Handle sampleLogSandboxHandle;    
     double now = WTF::currentTime();
-    String sampleLogFilePath = String::format("WebProcess%llu", static_cast<uint64_t>(now));
+    String sampleLogFilePath = String::format("WebProcess%llu", static_cast<unsigned long long>(now));
     sampleLogFilePath = SandboxExtension::createHandleForTemporaryFile(sampleLogFilePath, SandboxExtension::WriteOnly, sampleLogSandboxHandle);
     
     sendToAllProcesses(Messages::WebProcess::StartMemorySampler(sampleLogSandboxHandle, sampleLogFilePath, interval));
